refactor: QueryString 클래스를 common 패키지로 이동
