+++ src/main/java/webserver/common/QueryString.java
@@ -1,12 +1,12 @@
-package webserver.request;
+package webserver.common;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import webserver.request.exception.IllegalQueryStringException;
-import webserver.request.exception.IllegalQueryStringKeyException;
+import webserver.common.exception.IllegalQueryStringException;
+import webserver.common.exception.IllegalQueryStringKeyException;
 
 public class QueryString {
     private final Map<String, String> queryString;
