+++ src/main/java/webserver/common/exception/IllegalQueryStringException.java
@@ -1,4 +1,4 @@
-package webserver.request.exception;
+package webserver.common.exception;
 
 public class IllegalQueryStringException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 QueryString 입니다.";
