+++ src/main/java/webserver/common/exception/IllegalQueryStringKeyException.java
@@ -1,4 +1,4 @@
-package webserver.request.exception;
+package webserver.common.exception;
 
 public class IllegalQueryStringKeyException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 QueryString 의 키입니다.";
