+++ src/main/java/webserver/common/exception/IllegalSessionIdException.java
@@ -1,4 +1,4 @@
-package webserver.session.exception;
+package webserver.common.exception;
 
 public class IllegalSessionIdException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 sessionId 입니다.";
