+++ src/main/java/webserver/request/Uri.java
@@ -1,5 +1,7 @@
 package webserver.request;
 
+import webserver.common.QueryString;
+
 public class Uri {
     private final String path;
     private final QueryString queryString;
